# Speed up ugni's barrier by using chained transactions